+++ src/slic3r-shared/include/Slic3r/App/Scene/AabbRaycastNodeComponent.hpp
@@ -13,7 +13,7 @@ public:
         : m_aabb_mesh(mesh)
     {}
 
-    bool raycast(const Matrix4f& world, const Ray& ray, double& t) const override;
+    bool raycast(const Matrix4d& world, const Ray& ray, double& t) const override;
 
     Eigen::AlignedBox<float, 2> projected_bounding_box(
         const Matrix4f& mvp, const Slic3r::App::Render::Rect& viewport
