+++ src/slic3r-shared/include/Slic3r/App/Scene/CameraTrackballController.hpp
@@ -9,42 +9,42 @@ class CameraTrackballController
 public:
     explicit CameraTrackballController(Camera& camera) : m_camera(camera) {}
 
-    void set_focal_point(const Vec3f& pos)
+    void set_focal_point(const Vec3d& pos)
     {
         m_cam_focal = pos;
         update_camera();
     }
 
-    void set_focal_distance(float value)
+    void set_focal_distance(double value)
     {
         m_cam_focal_dist = std::max(MIN_FOCAL_DISTANCE, value);
         update_camera();
     }
 
-    void set_azimuth(float value)
+    void set_azimuth(double value)
     {
         m_azimuth = value;
         normalize_azimuth_and_zenith();
         update_camera();
     }
 
-    void set_zenith(float value);
+    void set_zenith(double value);
 
-    void add_azimuth(float value)
+    void add_azimuth(double value)
     {
         m_azimuth += value;
         normalize_azimuth_and_zenith();
         update_camera();
     }
 
-    void add_zenith(float value)
+    void add_zenith(double value)
     {
         m_zenith += value;
         normalize_azimuth_and_zenith();
         update_camera();
     }
 
-    void add_azimuth_and_zenith(float azimuth, float zenith)
+    void add_azimuth_and_zenith(double azimuth, double zenith)
     {
         m_azimuth += azimuth;
         m_zenith += zenith;
@@ -52,10 +52,10 @@ public:
         update_camera();
     }
 
-    const Vec3f& cam_focal() const { return m_cam_focal; }
-    float cam_focal_dist() const { return m_cam_focal_dist; }
-    float azimuth() const { return m_azimuth; }
-    float zenith() const { return m_zenith; }
+    const Vec3d& cam_focal() const { return m_cam_focal; }
+    double cam_focal_dist() const { return m_cam_focal_dist; }
+    double azimuth() const { return m_azimuth; }
+    double zenith() const { return m_zenith; }
 
 private:
     void update_camera();
@@ -63,12 +63,12 @@ private:
 private:
     Camera& m_camera;
 
-    constexpr static float MIN_FOCAL_DISTANCE = 1e-02;
+    constexpr static double MIN_FOCAL_DISTANCE = 1e-02;
 
-    Vec3f m_cam_focal{0,0,0};
-    float m_cam_focal_dist{30};
-    float m_azimuth{-M_PI_2};
-    float m_zenith{-M_PI_2};
+    Vec3d m_cam_focal{0,0,0};
+    double m_cam_focal_dist{30};
+    double m_azimuth{-M_PI_2};
+    double m_zenith{-M_PI_2};
 
 };
