+++ src/slic3r-shared/src/Slic3r/App/Scene/Plane.cpp
@@ -3,17 +3,17 @@
 
 namespace Slic3r::App::Scene {
 
-Plane Plane::from_point_and_vectors(const Vec3f& point, const Vec3f& v0, const Vec3f& v1)
+Plane Plane::from_point_and_vectors(const Vec3d& point, const Vec3d& v0, const Vec3d& v1)
 {
-    const Vec3f normal = v0.cross(v1);
-    const float d = -normal.dot(point);
+    const Vec3d normal = v0.cross(v1);
+    const double d = -normal.dot(point);
     return {normal, d};
 }
 
-bool Plane::intersects(const Ray& ray, float& t) const
+bool Plane::intersects(const Ray& ray, double& t) const
 {
-    const float denom = ray.direction.dot(normal);
-    if (std::fabs(denom) <= std::numeric_limits<float>::epsilon())
+    const double denom = ray.direction.dot(normal);
+    if (std::fabs(denom) <= std::numeric_limits<double>::epsilon())
         // the ray and plane are parallel => the only case there is no intersection
         return false;
